Move Functions to src/utility
