+++ src/config/constants.js
@@ -39,14 +39,3 @@ export const MAP_TRANSITION_DELAY = 500;
 export const MAX_ABILITY_SCORE = 20;
 // The default value of an ability score
 export const STARTING_ABILITY_SCORE_VALUE = 8;
-
-// Create the chest's name given it's (x, y) coordinates.
-export function chestName(map, x, y) {
-    return map + '(' + x + ',' + y + ')';
-}
-
-// Retrieve a coordinate (x, y) from a position.
-// This is mainly used to map the player to the specific (x, y) position in the map.
-export function spriteToCoordinates(position) {
-    return { x: position[0] / SPRITE_SIZE, y: position[1] / SPRITE_SIZE };
-}
