+++ src/features/player/actions/explore-chest.js
@@ -1,4 +1,4 @@
-import { chestName } from '../../../config/constants';
+import { getChestName } from '../../../utils/get-chest-name';
 
 export default function exploreChest(x, y) {
     return (dispatch, getState) => {
@@ -7,7 +7,7 @@ export default function exploreChest(x, y) {
         const { chests } = world;
 
         const { currentMap } = world;
-        const chest = chests[chestName(currentMap, x, y)];
+        const chest = chests[getChestName(currentMap, x, y)];
         if (chest !== undefined) {
             dispatch({
                 type: 'SET_CHEST_DATA',
