+++ src/utils/get-chest-name.js
@@ -1,0 +1,4 @@
+// Create the chest's name given it's (x, y) coordinates.
+export function getChestName(map, x, y) {
+    return map + '(' + x + ',' + y + ')';
+}
